+++ apps/precipitation/h60/drv_data_dynamic.py
@@ -155,31 +155,42 @@ class DrvData:
                 if os.path.exists(file_path_src_step):
 
                     # check compression mode
+                    compression_check = True
                     if self.compression_src:
                         file_path_tmp_step = remove_zip_extension(file_path_src_step)
-                        unzip_filename(file_path_src_step, file_path_tmp_step)
+                        compression_check = unzip_filename(file_path_src_step, file_path_tmp_step)
                     else:
                         file_path_tmp_step = deepcopy(file_path_src_step)
 
-                    # get dataset source
-                    obj_data_src = read_file_nc(file_path_tmp_step, file_variables=self.variables_src)
+                    # check unzipping process (if applied to the file)
+                    if compression_check:
 
-                    # organize variable(s) source
-                    vars_src_step = organize_data(obj_data_src, self.grid_geo_x_src, self.grid_geo_y_src)
+                        # get dataset source
+                        obj_data_src = read_file_nc(file_path_tmp_step, file_variables=self.variables_src)
 
-                    # save variable(s) obj to ancillary file
-                    folder_name_anc_raw_step, file_name_anc_raw_step = os.path.split(file_path_anc_raw_step)
-                    make_folder(folder_name_anc_raw_step)
-                    write_file_obj(file_path_anc_raw_step, vars_src_step)
+                        # organize variable(s) source
+                        vars_src_step = organize_data(obj_data_src, self.grid_geo_x_src, self.grid_geo_y_src)
+
+                        # save variable(s) obj to ancillary file
+                        folder_name_anc_raw_step, file_name_anc_raw_step = os.path.split(file_path_anc_raw_step)
+                        make_folder(folder_name_anc_raw_step)
+                        write_file_obj(file_path_anc_raw_step, vars_src_step)
+
+                        # info end time step
+                        alg_logger.info(' ----> Time Step "' + alg_time_step.strftime(time_format_algorithm) +
+                                        '" ... DONE')
+
+                    else:
+                        # info end time step (failed unzipping process)
+                        alg_logger.info(' ----> Time Step "' + alg_time_step.strftime(time_format_algorithm) +
+                                        '" ... FAILED. Unzip file "' + file_path_tmp_step +
+                                        '" failed for streams errors')
 
                     # delete uncompressed file (if needed)
                     if self.compression_src:
                         if os.path.exists(file_path_tmp_step):
                             os.remove(file_path_tmp_step)
 
-                    # info end time step
-                    alg_logger.info(' ----> Time Step "' + alg_time_step.strftime(time_format_algorithm) +
-                                    '" ... DONE')
                 else:
                     # info end time step
                     alg_logger.info(' ----> Time Step "' + alg_time_step.strftime(time_format_algorithm) +
@@ -232,7 +243,7 @@ class DrvData:
                     # info start adapt datasets
                     alg_logger.info(' -----> (2) Adapt datasets ... ')
                     obj_data_anc_adapt_step, obj_geo_x_adapt_step, obj_geo_y_adapt_step = adapt_data(
-                        obj_data_anc_raw_step,var_name_geo_x=geo_var_name_x, var_name_geo_y=geo_var_name_y)
+                        obj_data_anc_raw_step, var_name_geo_x=geo_var_name_x, var_name_geo_y=geo_var_name_y)
                     # info end adapt datasets
                     alg_logger.info(' -----> (2) Adapt datasets ... DONE')