+++ apps/soil_moisture/h122/drv_data_static.py
@@ -48,12 +48,18 @@ class DrvData:
             self.file_path_src = None
 
         # get geo grid destination information
-        self.geo_x_corner_ll_dst = self.alg_grid_destination['geo_x_corner_ll']
-        self.geo_x_corner_ur_dst = self.alg_grid_destination['geo_x_corner_ur']
-        self.geo_y_corner_ll_dst = self.alg_grid_destination['geo_y_corner_ll']
-        self.geo_y_corner_ur_dst = self.alg_grid_destination['geo_y_corner_ur']
-        self.geo_x_res_dst = self.alg_grid_destination['geo_x_res']
-        self.geo_y_res_dst = self.alg_grid_destination['geo_y_res']
+        if 'folder_name' in self.alg_grid_destination.keys() and 'file_name' in self.alg_grid_destination.keys():
+            self.folder_name_dst = self.alg_grid_destination['folder_name']
+            self.file_name_dst = self.alg_grid_destination['file_name']
+            self.grid_dst_type = 'file'
+        else:
+            self.geo_x_corner_ll_dst = self.alg_grid_destination['geo_x_corner_ll']
+            self.geo_x_corner_ur_dst = self.alg_grid_destination['geo_x_corner_ur']
+            self.geo_y_corner_ll_dst = self.alg_grid_destination['geo_y_corner_ll']
+            self.geo_y_corner_ur_dst = self.alg_grid_destination['geo_y_corner_ur']
+            self.geo_x_res_dst = self.alg_grid_destination['geo_x_res']
+            self.geo_y_res_dst = self.alg_grid_destination['geo_y_res']
+            self.grid_dst_type = 'geo'
 
     # method to organize data
     def organize_data(self):
@@ -68,12 +74,28 @@ class DrvData:
             geo_x_grid_src, geo_y_grid_src = None, None
 
         # define destination grid
-        geo_x_grid_dst, geo_y_grid_dst, geo_x_min_dst, geo_x_max_dst, geo_y_min_dst, geo_y_max_dst = create_grid_data(
-            self.geo_x_corner_ll_dst, self.geo_x_corner_ur_dst, self.geo_y_corner_ll_dst, self.geo_y_corner_ur_dst,
-            self.geo_x_res_dst, self.geo_y_res_dst)
+        if self.grid_dst_type == 'file':
+            geo_values_dst, geo_x_grid_dst, geo_y_grid_dst = read_grid_data(os.path.join(self.folder_name_dst, self.file_name_dst))
+
+            geo_x_min_dst, geo_x_max_dst = round(geo_x_grid_dst.min(), 7), round(geo_x_grid_dst.max(), 7)
+            geo_y_min_dst, geo_y_max_dst = round(geo_y_grid_dst.min(), 7), round(geo_y_grid_dst.max(), 7)
+
+            geo_values_dst[geo_values_dst < 0] = 0
+            geo_values_dst[geo_values_dst > 0] = 1
+
+        elif self.grid_dst_type == 'geo':
+            (geo_x_grid_dst, geo_y_grid_dst, geo_x_min_dst, geo_x_max_dst,
+             geo_y_min_dst, geo_y_max_dst) = create_grid_data(
+                self.geo_x_corner_ll_dst, self.geo_x_corner_ur_dst, self.geo_y_corner_ll_dst, self.geo_y_corner_ur_dst,
+                self.geo_x_res_dst, self.geo_y_res_dst)
+            geo_values_dst = None
+        else:
+            alg_logger.error(' ===> Destination grid type is not correctly defined')
+            raise NotImplementedError('Case not implemented yet')
 
         # organize grid obj
         grid_obj = {
+            'grid_geo_values_dst': geo_values_dst,
             'grid_geo_x_src': geo_x_grid_src, 'grid_geo_y_src': geo_y_grid_src,
             'grid_geo_x_dst': geo_x_grid_dst, 'grid_geo_y_dst': geo_y_grid_dst,
             'grid_geo_x_min_dst': geo_x_min_dst, 'grid_geo_x_max_dst': geo_x_max_dst,